+++ PowerEditor/src/ScintillaComponent/AutoCompletion.h
@@ -27,8 +27,8 @@ class ScintillaEditView;
 struct MatchedCharInserted {
 	MatchedCharInserted() = delete;
 	char _c;
-	int _pos;
-	MatchedCharInserted(char c, int pos) : _c(c), _pos(pos) {};
+	size_t _pos;
+	MatchedCharInserted(char c, size_t pos) : _c(c), _pos(pos) {};
 };
 
 class InsertedMatchedChars {
@@ -37,7 +37,7 @@ public:
 	void removeInvalidElements(MatchedCharInserted mci);
 	void add(MatchedCharInserted mci);
 	bool isEmpty() const { return _insertedMatchedChars.size() == 0; };
-	int search(char startChar, char endChar, int posToDetect);
+	intptr_t search(char startChar, char endChar, size_t posToDetect);
 
 private:
 	std::vector<MatchedCharInserted> _insertedMatchedChars;
